Fix Velocity import
